% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{add_paragraph_settings}
\alias{add_paragraph_settings}
\title{add settings for a paragraph}
\usage{
add_paragraph_settings(fp)
}
\arguments{
\item{fp}{formatting properties for paragraph, see \code{\link[officer]{fp_par}}.}
}
\description{
add settings for a paragraph into a Word document.
}
\examples{
library(officer)
add_paragraph_settings(fp_par(text.align = "center"))
}
