% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{block_pour_docx}
\alias{block_pour_docx}
\title{pour the content of a docx file}
\usage{
block_pour_docx(docx_file)
}
\arguments{
\item{docx_file}{external docx file path}
}
\description{
Pour the content of a docx file in the resulting docx
generated by the main R Markdown document.
}
