% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{block_toc}
\alias{block_toc}
\title{add a table of content}
\usage{
block_toc(level = 3, style = NULL, separator = ";")
}
\arguments{
\item{level}{max title level of the table}

\item{style}{optional. style in the document that will be used to build entries of the TOC.}

\item{separator}{optional. Some configurations need "," (i.e. from Canada) separator instead of ";"}
}
\description{
add a table of content as a block output in an R
Markdown document.
}
