% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borderfp.R
\name{borderfp}
\alias{borderfp}
\alias{format.borderfp}
\title{border properties object}
\usage{
borderfp(color = "black", style = "solid", width = 1)

\method{format}{borderfp}(x, type = "docx", side = "left", ...)
}
\arguments{
\item{color}{border color - single character value (e.g. "#000000" or "black")}

\item{style}{border style - single character value : "none" or "solid" or "dotted" or "dashed"}

\item{width}{border width - an integer value : 0>= value}

\item{x}{\code{borderfp} object}

\item{type}{output type}
}
\description{
create a border properties object.
}
\examples{
borderfp()
borderfp(color="orange", style="solid", width=1)
borderfp(color="gray", style="dotted", width=1)
}
