% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{chunk_column_break}
\alias{chunk_column_break}
\title{add a column break}
\usage{
chunk_column_break()
}
\description{
add a column break as a chunk output in an R
Markdown document. A column break is a code inserted in a
paragraph the tells Word where to end the current column
and begin the next.
}
\seealso{
\code{\link{sections}}
}
