% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{chunk_page_break}
\alias{chunk_page_break}
\title{add a page break}
\usage{
chunk_page_break()
}
\description{
add a page break as a chunk output in an R
Markdown document. A page break is a code inserted in a paragraph
that tells Word where to end the current page and begin the next.
}
