% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{chunk_seqfield}
\alias{chunk_seqfield}
\title{add a seqfield}
\usage{
chunk_seqfield(str, style)
}
\arguments{
\item{str}{code for seq field}

\item{style}{a character style name}
}
\description{
add a seqfield as a chunk output in an R
Markdown document. A seqfield is a placeholder for data
that can change, and it instructs Word to insert computed
values into a document automatically.
}
