% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{chunk_styled_text}
\alias{chunk_styled_text}
\title{styled text}
\usage{
chunk_styled_text(str, color = NULL, size = NULL, bold = FALSE,
  italic = FALSE, underlined = FALSE, font = NULL,
  valign = "baseline", shading = "transparent")
}
\arguments{
\item{str}{text}

\item{color}{font color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{size}{font size (in point) - 0 or positive integer value.}

\item{bold}{is bold}

\item{italic}{is italic}

\item{underlined}{is underlined}

\item{font}{single character value specifying font name.}

\item{valign}{single character value specifying font vertical alignments.
Expected value is one of the following : default \code{'baseline'}
or \code{'subscript'} or \code{'superscript'}}

\item{shading}{shading color - a single character value specifying
a valid color (e.g. "#000000" or "black").}
}
\description{
Character styles can be applied to individual chunk of text.
The function is creating a chunk of text associated with formatting
defined by several arguments.
}
\examples{
chunk_styled_text("test test", color = "red", bold = TRUE)
}
