% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{chunk_text_stylenamed}
\alias{chunk_text_stylenamed}
\title{add text and associate it with a character style name.}
\usage{
chunk_text_stylenamed(str, style)
}
\arguments{
\item{str}{chunk of text}

\item{style}{a character style name}
}
\description{
Character styles can be applied to individual chunk of text.
The function is creating a chunk of text associated with an existing
character style name located in the document used as base document.
}
