% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\title{Make A DrawingML plotting object to print in Powerpoint}
\usage{
dml(code, ggobj = NULL, layout = "default", bg = "white",
  fonts = list(), pointsize = 12, editable = TRUE)
}
\arguments{
\item{code}{plotting instructions}

\item{ggobj}{ggplot object to print. argument code will be ignored if this
argument is supplied.}

\item{layout}{either a character indicating layout type: "default",
"left_col", "right_col", "full_slide", or a list with elements "width",
"height", "left", and "top" in inches. Layouts are drawn from the
powerpoint template used for the R Markdown document.}

\item{bg, fonts, pointsize, editable}{Parameters passed to \code{\link{dml_pptx}}}
}
\description{
Only Powerpoint outputs currently supported. Pandoc >=2.4 is required.
Currently plots with rasters are not supported.
}
