% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{knit_print.dml}
\alias{knit_print.dml}
\title{Render a plot as a Powerpint DrawingML object}
\usage{
\method{knit_print}{dml}(x, ...)
}
\arguments{
\item{x}{a \code{dml} object}

\item{...}{further arguments, not used.}
}
\description{
Function used to render DrawingML in knitr/rmarkdown documents.
Only Powerpoint outputs currently supported
}
\author{
Noam Ross
}
