% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officedown.R
\docType{package}
\name{officedown}
\alias{officedown}
\title{officedown: 'R Markdown' Format for 'Microsoft' Documents}
\description{
The package \code{officedown} allows production of corporate documents from ‘R Markdown’ by reusing
formatting defined in ‘Microsoft Word’ documents. You can reuse table styles,
list styles but also add column sections, landscape oriented pages. Table and
image captions as well as cross-references are transformed into ‘Microsoft Word’
fields, allowing documents edition and merging without mis-referencing issues;
the syntax conforms to the ‘bookdown’ cross-reference definition. Objects
generated by the ‘officer’ package are also supported in the ‘knitr’ chunks.
}
\details{
‘Microsoft PowerPoint’ presentations also benefit from this as well as the
ability to produce editable vector graphics in PowerPoint.

To learn more about officer, start with the vignettes:
\code{browseVignettes(package = "officedown")}
}
\seealso{
\url{https://ardata-fr.github.io/officedown}
}
