% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocx_document.R
\name{rdocx_document2}
\alias{rdocx_document2}
\title{Wrapper for 'bookdown' documents}
\usage{
rdocx_document2(...)
}
\arguments{
\item{...}{Arguments to be passed to \link{rdocx_document} output format function.}
}
\description{
This is a simple 'bookdown' wrapper of the output format
\link{rdocx_document}
}
