% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{register_word_fig_caption}
\alias{register_word_fig_caption}
\title{knitr hook for figure caption autonumbering}
\usage{
register_word_fig_caption()
}
\description{
The function allows you to add a hook when executing
knitr to allow to turn figure captions into auto numbered figure
captions.

The user should not to call this function as it is called when the package
is attached, thus, running \code{library(officedown)} in your R Markdown
document should automatically register this hook.
}
