% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml.R
\name{sections}
\alias{sections}
\alias{block_section_continuous}
\alias{block_section_landscape}
\alias{block_section_portrait}
\alias{block_section_columns}
\title{section}
\usage{
block_section_continuous()

block_section_landscape(w = 21/2.54, h = 29.7/2.54)

block_section_portrait(w = 21/2.54, h = 29.7/2.54)

block_section_columns(widths = c(2.5, 2.5), space = 0.25,
  sep = FALSE)
}
\arguments{
\item{w, h}{width and height in inches of the section page. This will
be ignored if the default section (of the \code{reference_docx} file)
already has width and height.}

\item{widths}{columns widths in inches. If 3 values, 3 columns
will be produced.}

\item{space}{space in inches between columns.}

\item{sep}{if TRUE a line is separating columns.}
}
\description{
Add a section as a block output in an R
Markdown document. A section affects preceding paragraphs
or tables.
}
\details{
A section starts at the end of the previous section (or the beginning of
the document if no preceding section exists), and stops where the section is declared.
}
