% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textfp.R
\name{textfp}
\alias{textfp}
\title{Text formatting properties}
\usage{
textfp(color = "black", size = NULL, bold = FALSE, italic = FALSE,
  underlined = FALSE, font = NULL, vertical.align = "baseline",
  shading.color = "transparent")
}
\arguments{
\item{color}{font color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{bold}{is bold}

\item{italic}{is italic}

\item{underlined}{is underlined}

\item{vertical.align}{single character value specifying font vertical alignments.
Expected value is one of the following : default \code{'baseline'}
or \code{'subscript'} or \code{'superscript'}}

\item{shading.color}{shading color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{font.size}{font size (in point) - 0 or positive integer value.}

\item{font.family}{single character value specifying font name.}
}
\value{
a \code{textfp} object
}
\description{
Create a \code{textfp} object that describes
text formatting properties.
}
