% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocx_knit_print.R
\name{knit_print_block}
\alias{knit_print_block}
\title{Force block printing while knitting}
\usage{
knit_print_block(x, ...)
}
\arguments{
\item{x}{a block object, result of a block function from officer package}

\item{...}{unused arguments}
}
\value{
None. the function only print XML code.
}
\description{
When used in a loop, blocks output anything
because \code{knit_print} method is not called. In that situation,
use the function to force printing. Also you should tell the chunk
to use results 'as-is' (by adding \code{results='asis'} to your chunk header).
}
\section{Word example}{

Copy example located here:
\code{system.file(package = "officedown", "examples", "word_loop.Rmd")}
}

