% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocx_knit_print.R
\name{knit_print_run}
\alias{knit_print_run}
\title{Force run printing while knitting}
\usage{
knit_print_run(x, ...)
}
\arguments{
\item{x}{a run object, result of a run function from officer package}

\item{...}{unused arguments}
}
\description{
When used in a loop, runs output anything
because \code{knit_print} method is not called. In that situation,
use the function to force printing. Also you should tell the chunk
to use results 'as-is' (by adding \code{results='asis'} to your chunk header).
}
