% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocx_document.R
\name{rdocx_document}
\alias{rdocx_document}
\alias{rdocx_document2}
\title{Convert to an MS Word document}
\usage{
rdocx_document(mapstyles, base_format = rmarkdown::word_document, ...)

rdocx_document2(...)
}
\arguments{
\item{mapstyles}{a named list of style to be replaced in the generated
document. \code{list("Date"="Author")} will result in a document where
all paragraphs styled with stylename "Date" will be styled with
stylename "Author".}

\item{base_format}{format to be used as a base document for
officedown. default to \link[rmarkdown]{word_document} but
can also be word_document2 from bookdown}

\item{...}{arguments used by \link[rmarkdown]{word_document}}
}
\description{
Format for converting from R Markdown to an MS Word
document. The function comes also with improved output options.
\code{rdocx_document2} also supports cross reference based on the syntax of
the bookdown package.
}
\examples{
skeleton <- system.file(package = "officedown",
  "rmarkdown/templates/word/skeleton/skeleton.Rmd")
file.copy(skeleton, to = "officedown.Rmd")
library(rmarkdown)
render("officedown.Rmd", output_file = "officedown.docx")
}
