% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocx_document.R
\name{rdocx_document}
\alias{rdocx_document}
\title{Convert to an MS Word document}
\usage{
rdocx_document(mapstyles, ...)
}
\arguments{
\item{mapstyles}{a named list of style to be replaced in the generated
document. \code{list("Date"="Author")} will result in a document where
all paragraphs styled with stylename "Date" will be styled with
stylename "Author".}

\item{...}{arguments used by \link[rmarkdown]{word_document}}
}
\description{
Format for converting from R Markdown to an MS Word
document. The function comes also with improved output options.
}
\examples{
skeleton <- system.file(package = "worded",
  "rmarkdown/templates/worded/skeleton/skeleton.Rmd")
file.copy(skeleton, to = "worded.Rmd")
library(rmarkdown)
render("worded.Rmd", output_file = "worded.docx")
}
