% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpptx_document.R
\name{rpptx_document}
\alias{rpptx_document}
\title{Convert to an MS PowerPoint document}
\usage{
rpptx_document(...)
}
\arguments{
\item{...}{arguments used by \link[rmarkdown]{powerpoint_presentation}}
}
\description{
Format for converting from R Markdown to an MS PowerPoint
document.
}
